+++ src/LeaderboardsLayer.cpp
@@ -16,14 +16,12 @@ class $modify(LeaderboardsLayer) {
 	}
 
 	auto DayAndNightSystem = DayAndNightSystem::create();
-	DayAndNightSystem->setZOrder(-2);
 	DayAndNightSystem->setID("Events"_spr);
-	this->addChild(DayAndNightSystem);
+	this->addChild(DayAndNightSystem,-2);
 
 	auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-	DayAndNightSystemOverlay->setZOrder(106);
 	DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
-	this->addChild(DayAndNightSystemOverlay);
+	this->addChild(DayAndNightSystemOverlay,106);
 
 	
 		return true;