+++ src/MoreVideoOptionsLayer.cpp
@@ -12,9 +12,8 @@ class $modify(MoreVideoOptionsLayer) {
 
 
 	auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
-	DayAndNightSystemOverlay->setZOrder(106);
 	DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
-	this->addChild(DayAndNightSystemOverlay);
+	this->addChild(DayAndNightSystemOverlay,106);
 
 	
 		return true;